Update TorrentLeech (#4142)
* Update TorrentLeech

* Update Categories

* Add an info log message about results per page

* Replace reserved package name 'json'

* Move pagination handling into a function

* Change Results Per Page log to warning
